+++ Tiles/DriverConsole/Tiles/Bodies/IBody.cs
@@ -12,7 +12,7 @@ namespace Tiles.Bodies
         bool IsGrasping { get; }
         bool IsBeingGrasped { get; }
         bool IsWrestling { get; }
-        int Size { get; }
+        double Size { get; }
         IList<IBodyPart> Parts { get; }
 
         void Amputate(IBodyPart part);
