+++ Tiles/DriverConsole/Tiles/Items/ItemClass.cs
@@ -15,15 +15,15 @@ namespace Tiles.Items
         public IArmorClass ArmorClass { get; set; }
 
         public IMaterial Material { get; set; }
-        public int Size { get; set; }
-        
-        public ItemClass(string name, ISprite sprite, int size, IMaterial material, IWeaponClass weaponClass)
+        public double Size { get; set; }
+
+        public ItemClass(string name, ISprite sprite, double size, IMaterial material, IWeaponClass weaponClass)
             : this(name, sprite, size, material, weaponClass, null) { }
 
-        public ItemClass(string name, ISprite sprite, int size, IMaterial material, IArmorClass armorClass)
+        public ItemClass(string name, ISprite sprite, double size, IMaterial material, IArmorClass armorClass)
             : this(name, sprite, size, material, null, armorClass) { }
 
-        public ItemClass(string name, ISprite sprite, int size, IMaterial material, IWeaponClass weaponClass, IArmorClass armorClass)
+        public ItemClass(string name, ISprite sprite, double size, IMaterial material, IWeaponClass weaponClass, IArmorClass armorClass)
         {
             Name = name;
             Sprite = sprite;
